requirements.txt
aiohttp
alpaca-py
dash
dash-bootstrap-components
dotenv
numpy
pandas
plotly
python-dotenv
pytz
requests